fixed a couple uninitialized memory problems in the PPU and Rom. Hopefully this fixes the random crashes.

+++ cpp/mapper.cpp
@@ -32,11 +32,11 @@ UNROM::UNROM(Rom *rom) : Mapper(rom) {
 	this->rom = rom;
 }
 void UNROM::prg_write(word addr, byte val) {
-	cout << (int)rom << endl;
+	cout << (long)rom << endl;
 	cout << "switching banks " << int(val) << " " << (0x4000 * (val & 7)) << " " << (int)rom->prg_size << endl;
 	bank = val & 7;
 	rom->prg_rom[0] = rom->prg_banks + (0x4000 * (val & 7));
-	cout << (int)rom->prg_rom[0] << " " << (int)rom->prg_rom[0][0x3fff] << endl;
+	cout << (long)rom->prg_rom[0] << " " << (int)rom->prg_rom[0][0x3fff] << endl;
 };
 void UNROM::load_prg(byte prg_size, istream& f) {
 	rom->prg_rom = new byte*[2];
